Added user authentication

+++ app/Http/Controllers/TrackController.php
@@ -8,6 +8,7 @@ use App\Models\Tracked;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Auth;
 
 class TrackController extends Controller
 {
@@ -16,10 +17,9 @@ class TrackController extends Controller
         $request->validate([
             'domain' => 'required|string',
             'email' => 'required|email',
+            'expiry' => 'required|date',
             'days' => 'required|integer|min:1'
         ]);
-
-
 
         $domain = Str::of($request->input('domain'))
             ->replaceFirst('https://', '')
@@ -40,16 +40,23 @@ class TrackController extends Controller
         // Check if the domain is new (not already tracked)
         $existing = Tracked::where('domain', $domain)->first();
 
+        Tracked::create([
+            'domain'     => $request->domain,
+            'expiry'     => $request->expiry,
+            'email'      => $request->email,
+            'notifyDays' => $request->days,
+            'user_id'    => Auth::id(),
+        ]);
 
-        Tracked::updateOrCreate(
-            ['domain' => $domain],
-            [
+        // Tracked::updateOrCreate(
+        //     ['domain' => $domain],
+        //     [
 
-                'expiry' => $request->input('expiry'),
-                'email' => $request->input('email'),
-                'notifyDays' => (int) $request->input('days')
-            ]
-        );
+        //         'expiry' => $request->input('expiry'),
+        //         'email' => $request->input('email'),
+        //         'notifyDays' => (int) $request->input('days')
+        //     ]
+        // );
 
         // Log::info("Triggering domains:notify for newly tracked domain: {$domain}");
 
@@ -70,13 +77,13 @@ class TrackController extends Controller
 
     public function index()
     {
-        $tracked = Tracked::all();
+        $tracked = Tracked::where(('user_id'), Auth::id())->get();
         return view('tracked', compact('tracked'));
     }
 
     public function destroy($domain)
     {
-        Tracked::where('domain', $domain)->delete();
+        Tracked::where('domain', $domain)->where('user_id', Auth::id()) ->delete();
         return redirect()->route('track.index')->with('success', 'Domain removed.');
     }
 }
